MeeleWeapon

+++ Assets/_Scripts/ScriptableObjects/MeeleWeaponData.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewMeeleWeapon", menuName = "Weapons/Meele Weapon Data")]
 public class MeeleWeaponData : WeaponData
 {
-	public float durability;
+	public int durability;
 	public RuntimeAnimatorController animatiorController;
 }
